put categories and home components in pages folder
