+++ src/App.jsx
@@ -1,30 +1,8 @@
-import React, { useState } from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import React from 'react';
+import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 import NavigationComponent from './components/NavigationComponent';
-import BookListComponent from './components/BookListComponent';
-import BookFormComponent from './components/BookFormComponent';
-import Categories from './components/Categories';
-
-const Home = () => {
-  const [books, setBooks] = useState([]);
-
-  const handleAddBook = (newBook) => {
-    setBooks([...books, { ...newBook, id: Date.now() }]);
-  };
-
-  const handleDeleteBook = (bookId) => {
-    const updatedBooks = books.filter((book) => book.id !== bookId);
-    setBooks(updatedBooks);
-  };
-
-  return (
-    <div>
-      <h1>Home Page</h1>
-      <BookListComponent books={books} onDelete={handleDeleteBook} />
-      <BookFormComponent onAddBook={handleAddBook} />
-    </div>
-  );
-};
+import Categories from './pages/Categories';
+import Home from './pages/Home';
 
 const App = () => (
   <Router>
