+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as BookComponent } from './BookComponent';
+export { default as BookFormComponent } from './BookFormComponent';
+export { default as BookListComponent } from './BookListComponent';
+export { default as NavigationComponent } from './NavigationComponent';
